+++ components/ui/logo/spiderSpaceLogo.tsx
