+++ components/ui/techStackArea/techCard.tsx
