+++ components/ui/techStackArea/techStack.tsx
